@@ -30,8 +30,6 @@
     "mini-css-extract-plugin": "^0.4.1",
     "node-sass": "^4.9.0",
     "postcss-loader": "^2.1.5",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
     "react-excel-workbook": "0.0.4",
     "react-hot-loader": "^4.3.3",
     "react-redux": "^5.0.7",
@@ -51,6 +49,8 @@
   "dependencies": {
     "babel-polyfill": "^6.26.0",
     "lodash": "^4.17.10",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "webpack": "^4.12.0"
   }
 }
