@@ -1,17 +1,13 @@
-import React,{Component} from 'react';
-import context from 'components/context';
-const {Consumer}=context;
+import React, { Component } from 'react'
 
-export default class Child1 extends Component{
-  constructor(props){
-    super(props);
+export default class Child1 extends Component {
+  constructor(props) {
+    super(props)
   }
-  render(){
-    return(
+  render() {
+    return (
       <div>
-          <Consumer>
-            {value=><div style={{color:value}}>child1</div>}
-          </Consumer>
+        <div>child1</div>
       </div>
     )
   }
