@@ -1,8 +1,4 @@
-export const homeAjax={
-   url:"http://www.2-class.com/api/homePage/list",
-   method:'GET'
-}
-export const test={
-   url:"http://30.6.249.197:3010/api/peop/exam/getAnswerCalendar",
-   method:'GET'
+export const homeAjax = {
+  url: 'http://www.2-class.com/api/homePage/list',
+  method: 'GET'
 }
